Removed the elative (seestütlev) case completely.
Elative misleads: it was used in "bear left of ..." and the like;
one is still on the road when he/she is bearing there.

Yes, one is looking in the specified direction and sees things that are
away from the street. If the message were "look to the left of ...",
then elative would make sense, but not in our case.

Therefore, I prefer to use adessive (alalütlev) from now on, in such
cases.