Update default render; Addition colors for pedestrain roads in car and bicycle mode
In our default rendering profile we have for roads an additional="access=no" and an additional="access=private".
In such a case a road gets a slightly pinkish-red color instead of one of the normal colors.

Pedestrian roads are prohibited for cars and cyclists. IMO: a pedestrian road should get the same color in "car" and "bicycle" mode.