Update rendering_types; add missing rendering styles
Added a few rendering styles. The rendering styles are used in the routing, and work correctly as such, but they are not used in the rendering. The routing tells you you can't enter a street and that is correct) but the map doesn't tell you it is not allowed to enter a street: this is confusing.
(Thanks to user Allroads)