Add agricultural/forestry/delivery/private restrictions to "car" routing in routing.xml.
Cars should not be allowed into agricultural/forestry designated roads (both values are used according to taginfo and are documented). Also, going into private/delivery only roads should only be possible as last resort (e.g. to reach the final destination point).

Something should also be done with access=customers, but I am not sure what the 'value' should be in that case. It probably doesn't matter that there is no rule for that as such roads are often leaf and no exit roads so OsmAnd is not needing them for drive-though.