Rework "barriers" section in rounting.xml
In real routing, OsmAnd sometimes navigated me into a bollard. But there the passage is impossible by default. See http://wiki.openstreetmap.org/wiki/Tag:barrier%3Dbollard, the default is access only to foot and bicycle.

So I propose to rewrite the barriers section so that by default any barrier is impassable to a car. Only ones with access to a vehicle/car explicitly listed are passable. Lacking an access* tag, only make some known ones passable.

I am not sure what exactly the number in "value" means in this section. So it is possible I got some wrong. But I hope the idea is clear from the patch.