Update routing.xml; Refine penalties
Add this moment there are no penalties when going from motorway to motorway_link (and trunk to trunk_link). It sometimes happens that when a motorway_link exiting from a motorway, running parallel to the motorway, and leading back to the motorway is chosen by OsmAnd. This happens when the motorway_link is the inner bend and the motorway is the outer bend, causing the motorway_link to be 1-2 seconds faster in theory.
The other requirement is that both the motorway and motorway_link don't have a mxspeed or the same maxspeed.

Additionally I lowered the overall select value as in cities the value of 250 is a bit too much.

Solution for penalties contributed by Peter B in the mailing list.
Lowering overall value based on an issue in the mailing list and by a little trial & error by myself.