Observe "abutters" tag in routing.xml.
Some values of "abutters" mean that the road is inside a place (village, city, town), so a speed limit can apply. See http://wiki.openstreetmap.org/wiki/Abutters and https://wiki.openstreetmap.org/wiki/OSM_tags_for_routing/Maxspeed . It is not easily possible to determine this information from landuse or other methods (see the comments there), for some countries. For countries where it is possible is OsmAnd preprocessing the map files to include this information in some way?