Allow avoiding border crossings in car/bicycle routing.
Allow avoiding border crossings in car/bicycle routing. This can be useful for concave country parts (where the best route would be a shortcut trough another country). Or when there is a motorway near the border in the neighbouring country and OsmAnd would pick that as the fastest road. I have a real example of this case on a 400km trip (that OsmAnd calculated). It may be preferrable to avoid the foreign country, e.g. due to visa requirements, tolls, insurance coverage, etc. Of course best would be if OsmAnd considered administrative boundary relations to determine if a road is crossing into another country, but that is probably quite expensive for computation.