Add more weight options in routing.xml
I propose to add more weight options in routing.xml. Until there is a proper support for buses and hgvs in a special routing profile, I think these options need to be in car profile.

I wonder when HGV profile gets supported, it already can be enabled via developer plugin in OsmAnd. But there seem no special settings for it in routing.xml. I'd be happy if OsmAnd got a HGV and BUS profiles. Some of the maxheight, maxweight, maxlength and access options could be moved there so that the car profile does not have too many unnecessary options.