In routing, disallow lift_gate and swing_gate access for cars by default, but also support destination and customers access.
According to http://wiki.openstreetmap.org/wiki/Barriers barriers with undocumented access default to access=no. So if there is no explicit access tag on the node, consider lift_gate and swing_gate as impassable with a car too. Lift_gate wiki even mentions cars cannot pass by default.

I expect some paid/mall parking areas to be blocked by a lift_gate tagged with access=customers or access=destination. So let's allow passing those by car, of course with a very low priority, so that the road is only chosen if it is the "last mile" to the destination.