Proposals for bicycle routing - Part 1
Hello to everyone who is involved with Osmand development,

thank you for the new bicycle routing! The inclusion of elevation data is incredibly good. I am really impressed by how you implemented this! The graphic display was the biggest surprise - and works perfectly. 

I have some suggestions for further improvements in bicycle routing. This is the first part. Here, all the changes are below in the commentary (for this part, I have only made one minor change in the routing.xml: I have deleted an " ß " . This character leads to some android browsers being unable to display the file). 

Now, my first two suggestions:

1.) If a footway is marked as “sidewalk”, this is not recognisable in the Osmand maps at the moment. There is already a corresponding command in the routing.xml, but at the moment, it has no effect: <select value="0.3" t="footway" v="sidewalk"/> To make this command working, we have to include the “footway = sidewalk” data into the Osmand maps. Could someone implement this? (I do not know how to do it!). This would help to avoid many routing errors (for example in Vienna, where I have tested the new routing last week).

2.) There is still one bug concerning traffic signals. When a bicycle is running on a main road and approaches a big junction, it will often "hit upon" two tags for the traffic signals: First, there is a tag "highway=traffic_signal" (the traffic light for cars / bicycles on the road). Then, right behind, there is a tag "crossing=traffic_signals" (the traffic light for pedestrians crossing - sometimes this is just marked as "crossing=island", but not always). In Osmand, both tags have a "penalty" measured in seconds (the penalty is defined under <point attribute="obstacle">). BUT: When the bicycle is running on the main road, Osmand adds up both penalties, which is wrong: There should be only one penalty, because you do not have to wait two times when you are on the road (you just wait at the car traffic light, after it turns green, you pass the pedestria traffic light automatically). 

Again, unfortunately, I cannot implement the solution myself. The solution would be to change the way the maps are rendered: I there is a "crossing=traffic_signals" right behind a "highway=traffic_signals", Osmand could just erase the "highway=traffic_signals" tag in that spot. That tag is not needed in this special case. The "crossing=traffic_signals" is enough, because it already works in both directions: It imposes a penalty for anyone crossing the street AND for the cars / bicycles rolling on that street.

In ALL the other cases, we still need the two seperate penalties for "crossing=traffic_signals" AND for "highway=traffic_signals". These other cases are: a) A junction with a traffic light that is only marked with one single tag "highway=traffic_signals". B) A junction marked with "highway=traffic_signals" plus "crossing=island" for the pedestrians C) a traffic light without a junction (just for pedestrians to pass the road). That can sometimes be tagged with "crossing=traffic_signals" and "highway=traffic_signals" at the same time, but both tags are attached to the same node (I think in that case, Osmand would not add the two values, right?)

These two suggestions are very important, because without these two things being fixed, there are still some really weird errors in routing calculation.

Thank you!

Malte