refined values for elevation in bicycle mode
Dear all,

I have improved the values for the elevation mode and added them to the routing.xml. I did many hours of virtual testing in Vienna, Berlin and Hamburg. Before, the elevation mode sometimes would suggest a bad route, just to save 5 or 10 height meters. Now, it is more balanced (especially if we consider the "noise" that is in the elevation data). Also, now you get a clearer diffence between the three elevation modes (most leveled really suggesting the route with least elevation, etc).

I will do some further testing, but please do already include this into the routing.xml, because it is already a step forward.

If someone else wants to do some testing: A good starting point for a "calibration" are these two routes

a) From Marktgasse 21 (Vienna) to ORF-Zentrum (Kueniglberg, Venna). In safety mode on the last uphill section, the routing should take Wattmanngasse (residential) not the Maxingstrasse (tertiary). The routing over Maxingstrasse is supposed to have 7 height meters less (that could also just be noise in the data). Only to save 7 metres, it does not make sense to go 500 metres uphill on a tertiary road without a bike path (this only makes sense in "speed" mode).  If the routing takes Maxingstrasse in "safety" mode, this is a sign that the elevation gets too much weight compared to the other OSM data.

b) From the public swimming pool in Huetteldorf (Vienna) to Langenlebarn Volksschule. In "most leveled", the routing should make a detour that has an additional 13,5 kilometres, but at the same time saves 164 height meters (but not in "fast" mode). This route is good for checking the routing engine's preferences, because all three main alternatives have their shortcomings: The shortest route has a long part over secondary and tertiary road without a bike path. A slightly longer route is much safer, but has more elevation and a 4km section of track with unknown surface. And then there is the alternative that is very safe and very flat, but much longer. 

Greetings from Berlin,

Malte