some more changes for traffic lights
Please check also, if you like. 
I think these values solve some of the issues that we sometimes had with more complicated street crossings.

Again, this is a preliminary solution (we need a penalty for each and any crossing of a main road, and also for left turns on main roads).