correction for very steep downhill
According to xmd5a's comment, I have corrected the values for extremely steep downhill sections:

sections with -60 to -100 incline are not forbidden any more, the now only get a very high penalty. This is to make sure that we do not ban any ways that only have very high inclines due to inaccuracies of map data.