Added share_taxi=yes (#698)
The key share_taxi=yes is listed on https://wiki.openstreetmap.org/wiki/Key:psv and is in the wiki called "uncommon", beause it's not really a vehicle type, but it is used (actual over 22000 times) at public transport platforms to show that there stop shared taxis which are usually operated on demand. Because of the frequent usage, I added it so that people can see when this tag is used on a pt platform. That's also why it put it below the pt=platform category. It's not really a vehicle class, so it does not completely fit into the access category.