Update de_tts.js
Do not read out aloud ref= values for highway=secondary and highway=tertiary, if a street name is available.

Major roads highway=(motorway|primary|secondary|tertiary) in Germany 
have ref tags containing a type letter (A|B|L|K) and number, 
e.g. "ref=K 673" for a tertiary road. But only refs for motorways and highway=primary are
publicly visible on signposts, so reading the ref value of secondary or tertiary roads out loud
does not benefit the driver, as the information can not be cross-referenced with the
visible signs.