Consider cyclist and pedestrian hazards in routing
I recently learnt about the `hazard` tag. This adds some penalty for routing via a few types of hazardous places to bicycle and pedestrian routing. Cyclists and pedestrians are usually the parties negatively affected by these types of hazards. The penalty may in some cases result in hazards being avoided if an alternative route of otherwise similar quality is available.
It can be debated whether these hazards constitute "obstacle"s in the routing sense. But often for example a hazard=cyclists will mean merging onto a high-traffic road from a shoulder or kerb that may or may not be otherwise mapped and captured by routing, or hazard=dangerous_junction often will mean crossing a high-traffic road, both of which may require time to wait. Also, I only touched `obstacle` and left `obstacle_time` untouched.
Expansion of the routing file to include other hazards is of course possible, but this could be a start.